requirements.txt
python-telegram-bot~=13.11
pyyaml~=6.0
